requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==5.0.2
django-ckeditor-5==0.2.12
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-haystack @ git+https://github.com/django-haystack/django-haystack.git@77a8306ac7194f48082734b01e7209faeff43f90
django-simple-metatags==2.0.3
django-taggit==5.0.1
executing==2.0.1
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
mysqlclient==2.2.4
packaging==24.0
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
setuptools==69.2.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
tzdata==2024.1
Unidecode==1.3.8
wcwidth==0.2.13
Whoosh==2.7.4
